# docs/endpoints/openapi.yaml
openapi: 3.0.3
info:
  title: Observatorio Económico Core API
  version: 1.0.0
  description: Core estadístico para procedencia, metadatos, ingestión, revisión,
    consulta histórica, calidad y linaje.
servers:
- url: http://localhost:8080
  description: Local via NGINX
...
paths:
...
  /api/v1/provenance/organizations:
    post:
      operationId: createOrganization
      summary: Create producer or custodian organization
      tags:
      - Provenance
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - METHODOLOGY_STEWARD
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInput'
    get:
      operationId: listOrganizations
      summary: List organizations
      tags:
      - Provenance
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - ANALYST
      - METHODOLOGY_STEWARD
      security:
      - bearerAuth: []
      parameters: &id001
      - &id003
        name: page
        in: query
        schema:
          type: integer
          minimum: 1
          default: 1
      - &id004
        name: pageSize
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: search
        in: query
        schema:
          type: string
          maxLength: 120
  /api/v1/provenance/sources:
    post:
      operationId: createSource
      summary: Create statistical source
      tags:
      - Provenance
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - METHODOLOGY_STEWARD
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceInput'
    get:
      operationId: listSources
      summary: List sources
      tags:
      - Provenance
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - ANALYST
      - METHODOLOGY_STEWARD
      security:
      - bearerAuth: []
      parameters: *id001
  /api/v1/provenance/artifacts:
    post:
      operationId: registerSourceArtifact
      summary: Register immutable source artifact
      tags:
      - Provenance
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - DATA_OFFICER
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceArtifactInput'
  /api/v1/provenance/artifacts/{id}:
    get:
      operationId: getSourceArtifact
      summary: Get source artifact metadata
      tags:
      - Provenance
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - DATA_OFFICER
      - METHODOLOGY_STEWARD
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema: &id002
          type: string
          format: uuid
  /api/v1/data/observations:
    post:
      operationId: registerObservation
      summary: Register or revise one observation
      tags:
      - Data ingestion
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - DATA_OFFICER
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterObservationInput'
  /api/v1/data/observation-batches:
    post:
      operationId: importObservationBatch
      summary: Import up to 500 observations with partial results
      tags:
      - Data ingestion
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchImportResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-roles:
      - DATA_OFFICER
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationBatchInput'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          additionalProperties: false
          required:
          - code
          - message
        requestId:
          type: string
      additionalProperties: false
      required:
      - error
      - requestId
    OrganizationInput:
      type: object
      properties:
        parentOrganizationId: *id002
        code:
          type: string
          maxLength: 50
        legalName:
          type: string
        shortName:
          type: string
        organizationType:
          type: string
        countryCode:
          type: string
          pattern: ^[A-Z]{2}$
        officialStatisticsProducer:
          type: boolean
        isActive:
          type: boolean
        validFrom: &id005
          type: string
          format: date
        validTo: *id005
      additionalProperties: false
      required:
      - code
      - legalName
      - shortName
      - organizationType
      - countryCode
      - officialStatisticsProducer
      - validFrom
    SourceInput:
      type: object
      properties:
        organizationId: *id002
        frequencyId: *id002
        code:
          type: string
        name:
          type: string
        sourceType:
          type: string
        accessMethod:
          type: string
        officialUri:
          type: string
          format: uri
        licenseCode:
          type: string
        activeFrom: *id005
        activeTo: *id005
        isActive:
          type: boolean
      additionalProperties: false
      required:
      - organizationId
      - code
      - name
      - sourceType
      - accessMethod
    SourceArtifactInput:
      type: object
      properties:
        sourceId: *id002
        artifactType:
          type: string
        originalFilename:
          type: string
        originalUri:
          type: string
          format: uri
        storageUri:
          type: string
        mimeType:
          type: string
        sha256:
          type: string
          pattern: ^[a-f0-9]{64}$
        publicationDate: *id005
        retrievedAt:
          type: string
          format: date-time
        fileSizeBytes:
          type: string
          pattern: ^\d+$
        metadataJson:
          type: object
          additionalProperties: true
      additionalProperties: false
      required:
      - sourceId
      - artifactType
      - storageUri
      - sha256
      - retrievedAt
    DimensionValue:
      type: object
      properties:
        dimensionDefinitionId: *id002
        codeItemId: *id002
        classificationItemId: *id002
        geographicUnitId: *id002
        textValue:
          type: string
        numericValue:
          type: string
        dateValue: *id005
      additionalProperties: false
      required:
      - dimensionDefinitionId
    MeasureValue:
      type: object
      properties:
        measureDefinitionId: *id002
        numericValue:
          type: string
        textValue:
          type: string
        booleanValue:
          type: boolean
      additionalProperties: false
      required:
      - measureDefinitionId
    AttributeValue:
      type: object
      properties:
        attributeDefinitionId: *id002
        codeItemId: *id002
        numericValue:
          type: string
        textValue:
          type: string
        booleanValue:
          type: boolean
      additionalProperties: false
      required:
      - attributeDefinitionId
    ObservationRecord:
      type: object
      properties:
        indicatorVersionId: *id002
        frequencyId: *id002
        unitMeasureId: *id002
        periodStart: *id005
        periodEnd: *id005
        periodCode:
          type: string
        referenceDate: *id005
        dimensions:
          type: array
          items:
            $ref: '#/components/schemas/DimensionValue'
          maxItems: 50
        measures:
          type: array
          items:
            $ref: '#/components/schemas/MeasureValue'
          maxItems: 50
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/AttributeValue'
          maxItems: 50
        confidentialityStatus:
          type: string
        publicationDate: *id005
        vintageDate: *id005
        revisionReason:
          type: string
      additionalProperties: false
      required:
      - frequencyId
      - periodStart
      - periodEnd
      - periodCode
      - dimensions
      - measures
      - confidentialityStatus
      - vintageDate
    RegisterObservationInput:
      type: object
      properties:
        batchCode:
          type: string
          minLength: 3
          maxLength: 80
        datasetVersionId: *id002
        sourceArtifactId: *id002
        submittedByOrganizationId: *id002
        record:
          $ref: '#/components/schemas/ObservationRecord'
      additionalProperties: false
      required:
      - batchCode
      - datasetVersionId
      - sourceArtifactId
      - submittedByOrganizationId
      - record
    ObservationBatchInput:
      type: object
      properties:
        datasetVersionId: *id002
        sourceArtifactId: *id002
        submittedByOrganizationId: *id002
        batchCode:
          type: string
        entryMethod:
          type: string
          enum:
          - FILE
          - API
          - SDMX
          - DATABASE
          - MANUAL
        notes:
          type: string
        records:
          type: array
          items:
            $ref: '#/components/schemas/ObservationRecord'
          maxItems: 500
      additionalProperties: false
      required:
      - datasetVersionId
      - sourceArtifactId
      - submittedByOrganizationId
      - batchCode
      - entryMethod
      - records
    RegistrationResult:
      type: object
      properties:
        status:
          type: string
          enum:
          - PUBLISHED
          - REJECTED
          - UNCHANGED
        batchId: *id002
        seriesId: *id002
        observationId:
          type: string
          pattern: ^\d+$
        observationRevisionId:
          type: string
          pattern: ^\d+$
        revisionNumber:
          type: integer
          minimum: 1
        qualityIssueIds:
          type: array
          items: *id002
      additionalProperties: false
      required:
      - status
      - batchId
      - seriesId
      - observationId
      - qualityIssueIds
    BatchRecordResult:
      type: object
      properties:
        index:
          type: integer
          minimum: 0
        status:
          type: string
          enum:
          - PUBLISHED
          - REJECTED
          - UNCHANGED
          - INVALID
        observationId:
          type: string
          pattern: ^\d+$
        revisionId:
          type: string
          pattern: ^\d+$
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          additionalProperties: false
          required:
          - code
          - message
      additionalProperties: false
      required:
      - index
      - status
    BatchImportResult:
      type: object
      properties:
        batchId: *id002
        status:
          type: string
          enum:
          - COMMITTED
          - PARTIAL
          - REJECTED
        receivedCount:
          type: integer
          minimum: 1
        acceptedCount:
          type: integer
          minimum: 0
        rejectedCount:
          type: integer
          minimum: 0
        records:
          type: array
          items:
            $ref: '#/components/schemas/BatchRecordResult'
          maxItems: 500
      additionalProperties: false
      required:
      - batchId
      - status
      - receivedCount
      - acceptedCount
      - rejectedCount
      - records
...